move concatenate serialization tests to kernel tests
